+++ src/gql/types/CantoDamAssetType.php
@@ -5,9 +5,20 @@ namespace lsst\cantodamassets\gql\types;
 use craft\gql\base\ObjectType;
 use GraphQL\Type\Definition\ResolveInfo;
 use lsst\cantodamassets\gql\interfaces\CantoDamAssetInterface;
+use lsst\cantodamassets\lib\laravel\Collection;
+use yii\helpers\Inflector;
 
 class CantoDamAssetType extends ObjectType
 {
+    /**
+     * Which fields have their properties camelized to be compatible with GraphQL query params
+     */
+    private const CAMELIZED_FIELDS = [
+        'metadata',
+        'additional',
+        'default',
+    ];
+
     public function __construct(array $config)
     {
         $config['interfaces'] = [
@@ -20,6 +31,13 @@ class CantoDamAssetType extends ObjectType
     protected function resolve(mixed $source, array $arguments, mixed $context, ResolveInfo $resolveInfo): mixed
     {
         $fieldName = $resolveInfo->fieldName;
-        return $source[$fieldName] ?? null;
+        $resolvedData = $source[$fieldName];
+        // Make sure we camelize the keys if an array is being returned, since we normalize them to be camelized
+        // as GraphQL doesn't support spaces or other special characters in the query params
+        if (is_array($resolvedData) && in_array($fieldName, self::CAMELIZED_FIELDS, true)) {
+            $collection = new Collection($resolvedData);
+            $resolvedData = $collection->mapWithKeys(fn($value, $key) => [Inflector::camelize($key) => $value])->all();
+        }
+        return $resolvedData ?? null;
     }
 }